requirements.txt
Django>=4.2,<4.3
wagtail>=5.0,<5.1
django-extensions == 3.2.3
django-widget-tweaks == 1.4.12
sentry-sdk-1.25.1 ==1.25.1
